+++ lua_script_language.h
@@ -21,8 +21,9 @@
 
 #include "core/os/mutex.h"
 
-#include "lua_script.h"
 #include "parser/parser.h"
+
+#include "lua_script.h"
 
 class LuaScriptLanguage : public ScriptLanguage {
 	static LuaScriptLanguage *singleton;
@@ -44,6 +45,7 @@ private:
 private:
 	Mutex *mutex;
 	SelfList<LuaScript>::List script_list;
+	Parser parser{};
 
 public:
 	LuaScriptLanguage();
@@ -104,7 +106,7 @@ public:
 
 	virtual void get_recognized_extensions(List<String> *p_extensions) const;
 	virtual void get_public_functions(List<MethodInfo> *p_functions) const;
-	virtual void get_public_constants(List<Pair<String, Variant> > *p_constants) const;
+	virtual void get_public_constants(List<Pair<String, Variant>> *p_constants) const;
 
 	virtual void profiling_start();
 	virtual void profiling_stop();