+++ parser/ast/abstract_syntax_tree.cpp
@@ -19,6 +19,18 @@
 
 #include "abstract_syntax_tree.h"
 
-AbstractSyntaxTree::AbstractSyntaxTree() {}
+AbstractSyntaxTree::AbstractSyntaxTree(ast::Methods &&methods, bool valid) :
+		methods(methods),
+		valid(valid) {
+}
 
-AbstractSyntaxTree::~AbstractSyntaxTree() {}
+AbstractSyntaxTree::~AbstractSyntaxTree() {
+}
+
+bool AbstractSyntaxTree::is_valid() {
+	return this->valid;
+}
+
+ast::Methods AbstractSyntaxTree::get_methods() const {
+	return this->methods;
+}